folder structure and backendqueries
