+++ .eslintrc
@@ -23,6 +23,7 @@
     "no-multiple-empty-lines": ["error", { "max": 1 }],
     "quotes": ["error", "single", { "avoidEscape": true }],
     "object-curly-spacing": ["error", "always"],
+    "no-invalid-this": ["off"],
     "react/jsx-uses-vars": ["error"],
     "react/jsx-uses-react": ["error"]
   }
