+++ src/components/dynamic-field.css
@@ -6,7 +6,7 @@
   width: 100%;
   min-width: 100%;
   max-width: 100%;
-  padding: 3px;
+  padding: 2px 10px;
   word-break: break-word;
 }
 .dynamic_field[data-expanded='true'],
