+++ src/reducers.js
@@ -1,6 +1,6 @@
-import { transferObjectProps } from './util';
-import { transferQueryProps } from './util';
-import { copyObject } from './util';
+import { transferObjectProps } from './util/object.js';
+import { transferQueryProps } from './util/object.js';
+import { copyObject } from './util/object.js';
 
 // map previous global state to new global state based on action
 export function Reducer(prevState, action) {
@@ -86,47 +86,6 @@ export function Reducer(prevState, action) {
   if (!newState.pathQueries)
     newState.pathQueries = [];
 
-  // update url after state change unless on redux initialization or
-  // explicitly bypassed
-  if (action.updateUrl === true && prevState !== undefined)
-    updateUrl(newState);
-
   // set new state
   return newState;
-}
-
-// update url to reflect current state
-function updateUrl(state) {
-  // get checked metapaths
-  const checkedMetapaths = [];
-  for (const metapath of state.metapaths) {
-    if (metapath.checked)
-      checkedMetapaths.push(metapath.metapath_abbreviation);
-  }
-
-  checkedMetapaths.sort();
-
-  // new url
-  const newParams = new URLSearchParams();
-
-  // set url parameters
-  if (state.sourceNode.id !== undefined)
-    newParams.set('source', state.sourceNode.id);
-  if (state.targetNode.id !== undefined)
-    newParams.set('target', state.targetNode.id);
-  if (checkedMetapaths.length > 0)
-    newParams.set('metapaths', checkedMetapaths.join(','));
-
-  // make search string
-  let search = newParams.toString();
-  if (search.length > 0)
-    search = '?' + search;
-
-  // if url already matches state, dont push another history entry
-  if (search === window.location.search)
-    return;
-
-  // navigate to new url
-  const url = window.location.origin + window.location.pathname + search;
-  window.history.pushState({}, '', url);
 }
